+++ test/src/Shovel-test.js
@@ -1,0 +1,13 @@
+import { Shovel } from '../../src/Shovel.js';
+
+describe('Shovel', () => {
+  describe('ctor', () => {
+    it('throws if name is not passed', () => {
+      expect(() => new Shovel()).to.throw(TypeError).that.match(/name/);
+    });
+
+    it('throws if name is not a string', () => {
+      expect(() => new Shovel({})).to.throw(TypeError).that.match(/name/);
+    });
+  });
+});